refactor(shared): 분산된 LifecycleStatus 관련 타입 및 Enum 통합
- LifecycleStatus 타입과 Enum을 `shared` 모듈로 이동
- 관련 엔티티에서 `shared` 모듈 참조하도록 변경
- Next.js Content-Security-Policy 설정에서 불필요한 도메인 제거
