+++ src/shared/types/index.ts
@@ -1,1 +1,2 @@
 export * from './utility';
+export * from './domain';
